Move sqlite-specific funcs to sqlite_helper.hpp
